refactor:코드리뷰 수정(유틸 리팩토링)
